src/requirements.txt
requests==2.25.1
Django==3.2.2
Flask==1.0.3
mysql-connector-python==8.0.26
mysql-connector-python-rf==2.2.2
psycopg2==2.9.1
pip==19.2.3
psutil==5.8.0
PyYAML==5.3.1
drf-spectacular==0.19.0
uwsgi==2.0.19.1
djangorestframework==3.12.4